SpeedGrader slider view
refs:
affects:
release note:

test plan:
