+++ apps/teacher/src/main/java/com/instructure/teacher/fragments/SpeedGraderGradeFragment.kt
@@ -182,6 +182,8 @@ class SpeedGraderGradeFragment : BasePresenterFragment<SpeedGraderGradePresenter
 
         rubricEditView.setData(presenter.assignment, presenter.submission, presenter.assignee)
         rubricEditView.onAssessmentSaved = { presenter.updateSubmission(it) }
+
+        slider.setData(presenter.assignment, presenter.submission, presenter.assignee)
     }
 
     private fun showCustomizeGradeDialog() {
