+++ apps/teacher/src/main/java/com/instructure/teacher/view/grade_slider/SpeedGraderSliderEvents.kt
@@ -1,0 +1,8 @@
+package com.instructure.teacher.view.grade_slider
+
+import android.graphics.Rect
+import android.widget.SeekBar
+
+class ShowSliderGradeEvent(val seekBar: SeekBar?, val assigneeId: Long?, val description: String, val isTutorialTip: Boolean = false)
+
+class ShowLongHoldStartEvent()
